requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==0.3.14
langchain-groq==0.2.1
langchain-experimental==0.3.3
pandas==2.2.3
matplotlib==3.10.0
seaborn==0.13.2
streamlit==1.41.1
requests==2.32.3
python-dotenv==1.0.1
tabulate==0.9.0
